+++ Products/Reportek/Collection.py
@@ -29,6 +29,7 @@ import numbers
 import operator
 import os
 from datetime import datetime
+from decimal import Decimal, InvalidOperation
 
 import requests
 from AccessControl import ClassSecurityInfo, getSecurityManager
@@ -1543,7 +1544,15 @@ class Collection(CatalogAware, Folder, Toolz, DFlowCatalogAware, BaseCollection)
                 try:
                     if key in metadata:
                         amount = metadata.get(key, {}).get("Amount")
-                        mdata[key] = int(amount) if amount is not None else None
+                        try:
+                            mdata[key] = Decimal(amount) if amount is not None else None
+                        except (
+                            InvalidOperation,
+                            ValueError,
+                            KeyError,
+                            TypeError,
+                        ):
+                            mdata[key] = None
                     else:
                         mdata[key] = None
                 except (ValueError, KeyError, TypeError):